skanda_requirements.txt
# Skanda Purana PDF Extraction Requirements

# PDF Processing
PyPDF2>=3.0.1
PyMuPDF>=1.23.0

# OCR Capabilities
pytesseract>=0.3.10
Pillow>=9.5.0

# Indic Script Processing
indic-transliteration>=2.3.0

# Data Processing
pandas>=1.5.0
numpy>=1.24.0

# Text Processing
regex>=2023.0.0
unicodedata2>=15.0.0

# Progress Tracking
tqdm>=4.65.0

# Logging and Utilities
python-dateutil>=2.8.0